requirements.txt
celery 
redis
loguru
djangorestframework